+++ js/main.js
@@ -1,4 +1,4 @@
-import { loadState, saveState, DEFAULT_HERO_EMOJI, DEFAULT_DUNGEON_ENTRANCE_POSITION, migrateRingSlots, migrateBestDamage, migrateLoadout } from './state.js';
+import { loadState, saveState, DEFAULT_HERO_EMOJI, DEFAULT_DUNGEON_ENTRANCE_POSITION, migrateRingSlots, migratePowerRingSlot, migrateBestDamage, migrateLoadout } from './state.js';
 import { mountScreen, mountOverlay, unmountOverlay } from './screens/screenManager.js';
 import * as mapScreen from './screens/mapScreen.js';
 import * as battleScreen from './screens/battleScreen.js';
@@ -122,6 +122,7 @@ function startGame(loadedState, slotId) {
   state = migrateUpgradesToPerTier(loadedState);
   state = migrateNgPlusToolCarryover(state);
   state = migrateRingSlots(state);
+  state = migratePowerRingSlot(state);
   state = migrateBestDamage(state);
   state = migrateLoadout(state);
   activeSlotId = slotId;